+++ versions.tf
@@ -6,7 +6,7 @@ terraform {
   required_providers {
     aws = {
       source  = "hashicorp/aws"
-      version = ">= 3.29.0"
+      version = ">= 3.49.0"
     }
   }
 }