+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 2.0.0
+
+* maintenance (BREAKING): bump minimal aws provider version to 3.49.0
+* feat (BREAKING): Since provider v3.49.0, you can now use endpoint type "VPC". The service will take care of transfer familly endpoint creation
+* feat: add protocols support. You can now use SFTP, FTP and FTPS
+* feat: add certificate support for FTPS
+
 ## 1.2.0
 
 * feat: aws.vpc provider to support vpc endpoint creation with vpc peering subnets